eth/peer: refactor peerSet types and logic to own file
